backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
openai>=1.40.0
